backend/requirements.txt
Flask==3.1.*
flask-cors==5.0.*
mysqlclient==2.2.*
PyJWT==2.10.*
Werkzeug==3.1.*
python-dotenv==1.0.*
waitress==3.0.*
